Also use the same perf graph everywhere
