+++ examples/helpers/perf_graph.rs
@@ -1,0 +1,76 @@
+#![allow(unused)]
+
+use femtovg::{Align, Baseline, Canvas, Color, Paint, Path, Renderer};
+
+pub struct PerfGraph {
+    history_count: usize,
+    values: Vec<f32>,
+    head: usize,
+}
+
+impl PerfGraph {
+    pub fn new() -> Self {
+        Self {
+            history_count: 100,
+            values: vec![0.0; 100],
+            head: Default::default(),
+        }
+    }
+
+    pub fn update(&mut self, frame_time: f32) {
+        self.head = (self.head + 1) % self.history_count;
+        self.values[self.head] = frame_time;
+    }
+
+    pub fn get_average(&self) -> f32 {
+        self.values.iter().map(|v| *v).sum::<f32>() / self.history_count as f32
+    }
+
+    pub fn render<T: Renderer>(&self, canvas: &mut Canvas<T>, x: f32, y: f32) {
+        let avg = self.get_average();
+
+        let w = 200.0;
+        let h = 35.0;
+
+        let mut path = Path::new();
+        path.rect(x, y, w, h);
+        canvas.fill_path(&mut path, &Paint::color(Color::rgba(0, 0, 0, 128)));
+
+        let mut path = Path::new();
+        path.move_to(x, y + h);
+
+        for i in 0..self.history_count {
+            let mut v = 1.0 / (0.00001 + self.values[(self.head + i) % self.history_count]);
+            if v > 80.0 {
+                v = 80.0;
+            }
+            let vx = x + (i as f32 / (self.history_count - 1) as f32) * w;
+            let vy = y + h - ((v / 80.0) * h);
+            path.line_to(vx, vy);
+        }
+
+        path.line_to(x + w, y + h);
+        canvas.fill_path(&mut path, &Paint::color(Color::rgba(255, 192, 0, 128)));
+
+        let mut text_paint = Paint::color(Color::rgba(240, 240, 240, 255));
+        text_paint.set_font_size(12.0);
+        let _ = canvas.fill_text(x + 5.0, y + 13.0, "Frame time", &text_paint);
+
+        let mut text_paint = Paint::color(Color::rgba(240, 240, 240, 255));
+        text_paint.set_font_size(14.0);
+        text_paint.set_text_align(Align::Right);
+        text_paint.set_text_baseline(Baseline::Top);
+        let _ = canvas.fill_text(x + w - 5.0, y, &format!("{:.2} FPS", 1.0 / avg), &text_paint);
+
+        let mut text_paint = Paint::color(Color::rgba(240, 240, 240, 200));
+        text_paint.set_font_size(12.0);
+        text_paint.set_text_align(Align::Right);
+        text_paint.set_text_baseline(Baseline::Alphabetic);
+        let _ = canvas.fill_text(
+            x + w - 5.0,
+            y + h - 5.0,
+            &format!("{:.2} ms", avg * 1000.0),
+            &text_paint,
+        );
+    }
+}
